+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 - Added DatePicker component
 - Added DateRangePicker component
 - Added DateTimePicker component
+- Added TimePicker component
 
 ## 1.0.0 - 06-09-2020